Create Home Automation.ino
This arduino code helps to control a sony tv with a small IOT device with an IR transmitting led.
I have developed an android app and included some buttons to check the working of this code. It worked perfectly. If you want to use my code you have to do some changes like  wifi settings and firebase settings. You may be able to operate the tv from any part of the world. you don't need to stand near the tv.